Enusre public header and generated code have no implicit converion.

+++ src/google/protobuf/map_entry_lite.h
@@ -250,11 +250,11 @@ class MapEntryImpl : public Base {
   int GetCachedSize() const {
     int size = 0;
     size += has_key()
-        ? kTagSize + KeyTypeHandler::GetCachedSize(key())
+        ? static_cast<int>(kTagSize + KeyTypeHandler::GetCachedSize(key()))
         : 0;
     size += has_value()
-        ? kTagSize + ValueTypeHandler::GetCachedSize(
-            value())
+        ? static_cast<int>(kTagSize + ValueTypeHandler::GetCachedSize(
+            value()))
         : 0;
     return size;
   }
